changed double to int

+++ edu/jalc/watch/interfaces/Notifiable.java
@@ -1,5 +1,5 @@
 package edu.jalc.watch.interfaces;
 
 public interface Notifiable{
-  public void notified(double time);
+  public void notified(int time);
 }
